Render aoi, charts_data and dataset in streamlit frontend
