+++ src/frontend/pages/2_🧵_Threads.py
@@ -3,6 +3,8 @@ import json
 import requests
 import streamlit as st
 from app import API_BASE_URL
+from utils import render_stream
+from client import ZenoClient
 
 st.set_page_config(page_title="🧵 Threads", page_icon="🧵")
 
@@ -18,30 +20,16 @@ def get_auth_headers():
 
 # Fetch threads
 def fetch_threads():
-    resp = requests.get(
-        f"{API_BASE_URL}/api/threads", headers=get_auth_headers()
-    )
-
-    if resp.status_code == 200:
-        return resp.json()
-
-    st.error("Failed to fetch threads")
-    return []
+    client = ZenoClient(base_url=API_BASE_URL, token=st.session_state.token)
+    return client.list_threads() if st.session_state.get("token") else []
 
 
 # Fetch a single thread
 def fetch_thread(thread_id):
-    with requests.get(
-        f"{API_BASE_URL}/api/threads/{thread_id}",
-        headers=get_auth_headers(),
-        stream=True,
-    ) as stream:
-        st.subheader(f"Thread {selected_id}")
+    client = ZenoClient(base_url=API_BASE_URL, token=st.session_state.token)
 
-        for chunk in stream.iter_lines():
-            if chunk:
-                data = json.loads(chunk.decode("utf-8"))
-                st.write(data)
+    for stream in client.fetch(thread_id):
+        render_stream(stream)
 
 
 threads = []
